+++ ext/pseudocode.jl
@@ -1,0 +1,4 @@
+config = InversionConfig()
+data = InversionData()
+reconstruct_params!(fip, x)
+prob = InversionProblem(config, data, reconstruct_params)
