+++ src/FastIsostasy.jl
@@ -3,17 +3,8 @@ module FastIsostasy
 using AbstractFFTs: AbstractFFTs
 using CUDA: CuArray, CuMatrix, CUFFT, allowscalar
 using DelimitedFiles: readdlm
-using Distributions: MvNormal
 using Downloads: download
 using DynamicalSystemsBase: CoupledODEs, trajectory
-using EnsembleKalmanProcesses: EnsembleKalmanProcesses, EnsembleKalmanProcess,
-                               Unscented, get_error, get_g_mean_final,
-                               get_ϕ_final, get_ϕ_mean_final
-using EnsembleKalmanProcesses.Observations: Observations
-using EnsembleKalmanProcesses.ParameterDistributions: ParameterDistributions,
-                                                      ParameterDistribution,
-                                                      combine_distributions,
-                                                      constrained_gaussian
 using FastGaussQuadrature: gausslegendre
 using FFTW: fft, ifft, plan_fft!, plan_ifft!, cFFTWPlan
 using Interpolations: Flat, Gridded, Linear, OnGrid, Throw, linear_interpolation
@@ -26,16 +17,13 @@ using ParallelStencil: ParallelStencil, @init_parallel_stencil, @parallel,
                        @parallel_indices
 using Statistics: mean, cov
 using SpecialFunctions: besselj0, besselj1
-using .Threads
+# using .Threads
 
 using Reexport: Reexport, @reexport
 @reexport using Interpolations
 @reexport using OrdinaryDiffEq: Euler, Midpoint, Heun, Ralston, BS3, BS5, RK4,
     OwrenZen3, OwrenZen4, OwrenZen5, Tsit5, DP5, RKO65, TanYam7, DP8,
     Feagin10, Feagin12, Feagin14, TsitPap8, Vern6, Vern7, Vern8, Vern9
-    # SSPRK22, SSPRK33, SSPRK53, SSPRK63, SSPRK73, SSPRK83, SSPRK432, SSPRK43,
-    # SSPRK932, SSPRK54, SSPRK104, SSPRKMSVS32, SSPRKMSVS43, SSPRK53_2N1, SSPRK53_2N2,
-    # KenCarp47
 
 include("convenience_types.jl")
 include("convolution.jl")
@@ -47,10 +35,10 @@ include("derivatives_parallel.jl")
 include("integrators.jl")
 include("geostate.jl")
 include("mechanics.jl")
-include("inversion.jl")
 include("analytic_solutions.jl")
 include("dataloaders.jl")
 include("elra.jl")
+include("inversion.jl")
 
 # convolution.jl
 export InplaceConvolution
@@ -89,10 +77,6 @@ export build_greenintegrand, get_quad_coeffs, get_elasticgreen, get_geoidgreen
 # integrators.jl
 export simple_euler!, SimpleEuler
 
-# inversion.jl
-export InversionConfig, InversionData, InversionProblem, solve, extract_inversion
-export get_ϕ_mean_final     # from EnsembleKalmanProcesses.jl
-
 # analytic solutions
 export analytic_solution
 
@@ -103,6 +87,9 @@ export load_lithothickness_pan2022, load_logvisc_pan2022
 export load_ice6gd
 export load_spada2011, spada_cases
 export load_latychev_test3, load_latychev2023_ICE6G
+
+# inversion.jl
+export InversionConfig, InversionData, InversionProblem
 
 export update_second_derivatives!
